Reformat package
