--- src/main/java/historyObjects/Festival.java
@@ -1,53 +1,0 @@
-package historyObjects;
-
-import com.google.gson.annotations.SerializedName;
-
-public class Festival {
-
-	@SerializedName("Lễ hội truyền thống")
-	private String festivalName;
-	
-	@SerializedName("Ngày âm lịch")
-	private String lunarCalendarDate;
-	
-	@SerializedName("Ghi chú")
-	private String note;
-	
-	@SerializedName("Lần đầu tổ chức năm")
-	private String theFirstTime_Year;
-	
-	@SerializedName("Vị trí")
-	private String location;
-
-	public String getFestivalName() {
-		return festivalName;
-	}
-
-	public String getLunarCalendarDate() {
-		return lunarCalendarDate;
-	}
-
-	public String getNote() {
-		return note;
-	}
-
-	public String getTheFirstTime_Year() {
-		return theFirstTime_Year;
-	}
-
-	public String getLocation() {
-		return location;
-	}
-
-	@Override
-	public String toString() {
-		return "Festival{"
-				+ "festivalName='" + festivalName + "\'"
-				+ ", lunarCalendarDate='" + lunarCalendarDate + "\'"
-				+ ", note='" + note + "\'"
-				+ ", theFirstTime_Year='" + theFirstTime_Year + "\'"
-				+ ", location=" + location + "\'"
-			+ "}";
-	}
-	
-}
