+++ src/main/java/historyobject/HistoryObject.java
@@ -5,7 +5,7 @@ public class HistoryObject {
     private static final String DEFAULT_STRING_VALUE = "không có";
 
     // Lazy loading: return default value if field is null
-    public static String lazyLoading(String value){
-        return (value == null)? DEFAULT_STRING_VALUE : value;
+    public static String lazyLoading(String value) {
+        return (value == null) ? DEFAULT_STRING_VALUE : value;
     }
 }
